Trying to fix issues with io-structures used for ethercat interface.
Two thoughts, first I am going to try to stop modifying the struct used for beam parameter exchange (subsys<->arbiter), and second I created two structs, one for use within the library code and one for IO (called *_IO). Hopefully doing this will avoid other libraries from retaining old versions of a struct that really should only come from PMPS.

The type system in TC is confusing and enigmatic as it is vital so I want to mess with it as little as possible.