Veto arbiter work.
eV tests are better now, more agnostic to different eV ranges, more focused on the logic.

Also committing some tests of override logic in fast faults. I'm a bit mystified by an ongoing issue with the PEW on the L line. It says it's faulted but doesn't trip the beam off....