Another commit for the record.
This one shows that the arbiter is functional with a mismatch between the dataPool size and the entries array size. By simply tripling the size of the entry array to be 3x the size of the data pool the whole thing works and EPICS will work without additional copying.

However I consider this to be unsafe as a future change may cause an out of bounds access issue. Thankfully the dataPool is loaded sequentially without skipping so a memcpy will be enough for the EPICS display array.